feat: 포스터가 없는 것은 다른 이미지로 대체
