+++ BoxOffice/Model/BoxOfficeModel/DailyBoxOffice.swift
@@ -1,0 +1,12 @@
+//
+//  DayBoxOffice.swift
+//  BoxOffice
+//
+//  Created by 박도원 on 2023/01/03.
+//
+
+import Foundation
+
+struct DailyBoxOffice: Decodable, Hashable {
+    let boxOfficeResult: BoxOfficeResult
+}
