+++ BoxOffice/Model/FilmDetailModel/FilmDetails.swift
@@ -1,0 +1,12 @@
+//
+//  FilmDetails.swift
+//  BoxOffice
+//
+//  Created by 천승희 on 2023/01/03.
+//
+
+import Foundation
+
+struct FilmDetails: Decodable {
+    let movieInfoResult: MovieInfoResult
+}
