+++ BoxOffice/Model/FilmDetailModel/MovieInfoResult.swift
@@ -1,0 +1,12 @@
+//
+//  File.swift
+//  BoxOffice
+//
+//  Created by 박도원 on 2023/01/06.
+//
+
+import Foundation
+
+struct MovieInfoResult: Decodable {
+    let movieInfo: MovieInfo
+}
