+++ BoxOffice/Model/FilmDetailModel/PeopleName.swift
