+++ BoxOffice/Model/Manager/NetworkManager.swift
@@ -15,8 +15,6 @@ class NetworkManager: NetworkProtocol {
         let url = "https://kobis.or.kr/kobisopenapi/webservice/rest/boxoffice/searchDailyBoxOfficeList.json?key=\(appKey)&targetDt=\(currentDate)"
         self.getData(url: url, completion: completion)
     }
-    
-    
 }
 
 extension NetworkManager {
